requirements.txt
streamlit==0.82.0
numpy==1.19.5
tensorflow==2.5.0
tensorflow-hub==0.12.0
matplotlib==3.4.2
opencv-python
Pillow==8.2.0
